Angularize UI

+++ autosubliminal/core/cache.py
@@ -3,7 +3,6 @@
 import datetime
 import logging
 import os
-import shutil
 import tempfile
 
 import requests
@@ -13,7 +12,6 @@ from dogpile.util.readwrite_lock import ReadWriteMutex
 from six import text_type
 
 import autosubliminal
-from autosubliminal.util.common import set_rw_and_remove
 
 log = logging.getLogger(__name__)
 
@@ -43,17 +41,13 @@ class MutexFileLock(AbstractFileLock):
         return self.mutex.release_write_lock()
 
 
-def clear_mako_cache():
-    mako_cache_dir = os.path.abspath(os.path.join(autosubliminal.CACHEDIR, 'mako'))
-    shutil.rmtree(mako_cache_dir, onerror=set_rw_and_remove)
-
-
 def clear_imdbpie_cache():
     # Cache is created by imdbpie in temp location (see auth.py in imdbpie)
     # Cleanup is required when switching between python versions
     # If not, 'ValueError: unsupported pickle protocol' is thrown
     imdb_cache_file = os.path.abspath(os.path.join(tempfile.gettempdir(), 'cache.db'))
-    os.remove(imdb_cache_file)
+    if os.path.exists(imdb_cache_file):
+        os.remove(imdb_cache_file)
 
 
 def cache_artwork(indexer_name, indexer_id, artwork_type, artwork_url, thumbnail=False):
